+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/HealthProfHome.java
@@ -19,15 +19,13 @@ import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
 import com.triadss.doctrack2.R;
 import com.triadss.doctrack2.activity.LoginActivity;
-import com.triadss.doctrack2.activity.healthprof.fragment.HealthProfHomeFragment;
-import com.triadss.doctrack2.activity.healthprof.fragment.HealthProfessionalAppointmentFragment;
-import com.triadss.doctrack2.activity.healthprof.fragment.HealthProfessionalReportFragment;
-import com.triadss.doctrack2.activity.patient.fragment.PatientHomeFragment;
+import com.triadss.doctrack2.activity.healthprof.fragments.HealthProfHomeFragment;
+import com.triadss.doctrack2.activity.healthprof.fragments.HealthProfessionalAppointmentFragment;
+import com.triadss.doctrack2.activity.healthprof.fragments.HealthProfessionalReportFragment;
 import com.triadss.doctrack2.config.constants.NotificationConstants;
 import com.triadss.doctrack2.databinding.ActivityHealthProfHomeBinding;
-import com.triadss.doctrack2.activity.healthprof.fragment.AppointmentFragment;
 import com.triadss.doctrack2.activity.core.DeviceFragment;
-import com.triadss.doctrack2.activity.healthprof.fragment.PatientFragment;
+import com.triadss.doctrack2.activity.healthprof.fragments.PatientFragment;
 import com.triadss.doctrack2.activity.patient.fragment.RecordFragment;
 import com.triadss.doctrack2.notification.NotificationService;
