+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/HealthProfHomeFragment.java
@@ -1,6 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof.fragment;
-
-import static androidx.constraintlayout.helper.widget.MotionEffect.TAG;
+package com.triadss.doctrack2.activity.healthprof.fragments;
 
 import android.content.Context;
 import android.content.Intent;
@@ -11,13 +9,11 @@ import androidx.fragment.app.Fragment;
 import androidx.recyclerview.widget.LinearLayoutManager;
 import androidx.recyclerview.widget.RecyclerView;
 
-import android.util.Log;
 import android.view.LayoutInflater;
 import android.view.View;
 import android.view.ViewGroup;
 import android.widget.Button;
 import android.widget.TextView;
-import android.widget.Toast;
 
 import com.google.firebase.auth.FirebaseAuth;
 import com.google.firebase.auth.FirebaseUser;
@@ -26,7 +22,6 @@ import com.triadss.doctrack2.activity.LoginActivity;
 import com.triadss.doctrack2.activity.patient.fragment.PatientHomeAppointmentAdapter;
 import com.triadss.doctrack2.config.constants.SessionConstants;
 import com.triadss.doctrack2.dto.AppointmentDto;
-import com.triadss.doctrack2.dto.DateTimeDto;
 import com.triadss.doctrack2.repoositories.AppointmentRepository;
 import com.triadss.doctrack2.repoositories.ReportsRepository;
