+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/HealthProfessionalAppointmentStatusAdapter.java
@@ -1,4 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof.fragment;
+package com.triadss.doctrack2.activity.healthprof.fragments;
 
 import android.content.Context;
 import android.view.LayoutInflater;
