+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/ShowPendingAppointment.java
@@ -1,4 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof.fragment;
+package com.triadss.doctrack2.activity.healthprof.fragments;
 
 import android.os.Bundle;
