+++ app/src/main/java/com/triadss/doctrack2/activity/healthprof/fragments/UpdateMedicalHistory.java
@@ -1,4 +1,4 @@
-package com.triadss.doctrack2.activity.healthprof.fragment;
+package com.triadss.doctrack2.activity.healthprof.fragments;
 
 import android.content.Context;
 import android.content.SharedPreferences;
