+++ app/src/main/res/layout/fragment_appointment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.healthprof.fragment.AppointmentFragment">
+    tools:context=".activity.healthprof.fragments.AppointmentFragment">
 
     <!-- TODO: Update blank fragment layout -->
     <TextView
