+++ app/src/main/res/layout/fragment_health_professional_appointment.xml
@@ -3,7 +3,7 @@
     xmlns:tools="http://schemas.android.com/tools"
     android:layout_width="match_parent"
     android:layout_height="match_parent"
-    tools:context=".activity.healthprof.fragment.HealthProfessionalAppointmentFragment">
+    tools:context=".activity.healthprof.fragments.HealthProfessionalAppointmentFragment">
 
     <LinearLayout
         android:layout_width="match_parent"
